Added: parseAssign validations to check if correct json being passed
