+++ validations/utils.js
@@ -1,0 +1,8 @@
+function getTemplates(engine, expression) {
+  const templates = engine.parse(expression);
+  return templates;
+}
+
+module.exports = {
+  getTemplates,
+};
